+++ src/interfaces/IDistributionStrategy.sol
@@ -10,7 +10,7 @@ interface IDistributionStrategy {
     /// @param distributionContract The contract that was created that will handle or manage the distribution.
     /// @param token The token that is being distributed.
     /// @param totalSupply The supply of the token that is being distributed.
-    event DistributionInitialized(address indexed distributionContract, address indexed token, uint128 totalSupply);
+    event DistributionInitialized(address indexed distributionContract, address indexed token, uint256 totalSupply);
 
     /// @notice Initialize a distribution of tokens under this strategy.
     /// @dev Contracts can choose to deploy an instance with a factory-model or handle all distributions within the
@@ -22,7 +22,7 @@ interface IDistributionStrategy {
     /// @param salt The optional salt for deterministic deployment.
     /// @return distributionContract The contract that will handle or manage the distribution.
     ///         (Could be `address(this)` if the strategy is handled in-place, or a newly deployed instance).
-    function initializeDistribution(address token, uint128 totalSupply, bytes calldata configData, bytes32 salt)
+    function initializeDistribution(address token, uint256 totalSupply, bytes calldata configData, bytes32 salt)
         external
         returns (IDistributionContract distributionContract);
 }
